tag chips now draggable both directions.
mobile click version of moving them almost done.

lastly not sure how the json is supposed to be set
up, is it one json file per question. one json
object with different answer choices based on an
ID, or do they all share the same json file for
the page?...